# lib/api-spec/openapi.orval.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api",
    "version": "0.2.0",
    "description": "Docuvia — Universal VCS Knowledge Graph Engine API"
  },
  "servers": [{ "url": "/api", "description": "Base API path" }],
...
  "paths": {
...
    "/projects/{id}/ingest/git": {
      "post": {
        "operationId": "ingestGit",
        "tags": ["ingest"],
        "summary": "Ingest commits from a Git repository (GitHub API)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GitIngestInput" } }
          }
        },
        "responses": {
          "200": {
            "description": "Ingest result",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GitIngestResult" } }
            }
          }
        }
      }
    },
    "/projects/{id}/ingest/status": {
      "get": {
        "operationId": "getProjectIngestStatus",
        "tags": ["ingest"],
        "summary": "Get ingest status and cursor info for a project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Ingest status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IngestStatusResponse" }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/ingest/svn": {
      "post": {
        "operationId": "ingestSvn",
        "tags": ["ingest"],
        "summary": "Ingest SVN commits into the knowledge graph",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SvnIngestInput" } }
          }
        },
        "responses": {
          "200": {
            "description": "Ingest result",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SvnIngestResult" } }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/projects/{id}/ingest/document": {
      "post": {
        "operationId": "ingestDocument",
        "tags": ["ingest"],
        "summary": "Ingest a document (MD, TXT) into the project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DocumentIngestInput" } }
          }
        },
        "responses": {
          "201": {
            "description": "Stored document",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Document" } }
            }
          }
        }
      }
    },
    "/projects/{id}/ingest/document/upload": {
      "post": {
        "operationId": "uploadDocument",
        "tags": ["ingest"],
        "summary": "Upload and parse a document file (PDF, DOCX, PPTX, MD)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/DocumentUploadInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded and parsed document record",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Document" } }
            }
          },
          "400": {
            "description": "Unsupported file type",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "413": {
            "description": "File too large (exceeds 10MB limit)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/projects/{id}/documents": {
      "get": {
        "operationId": "listDocuments",
        "tags": ["documents"],
        "summary": "List documents for a project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Document" } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
      "DocumentUploadInput": {
        "type": "object",
        "required": ["file", "l2NodeId"],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The document file to upload (max 10MB)"
          },
          "l2NodeId": { "type": "integer", "description": "Required L2 Module ID" },
          "docType": {
            "type": "string",
            "enum": ["pdf", "docx", "pptx", "md"],
            "description": "Document type (auto-detected from extension if omitted)"
          },
          "commitSha": {
            "type": "string",
            "description": "Optional commit SHA to associate with this document"
          }
        }
      },
      "Document": {
        "type": "object",
        "required": ["id", "l2NodeId", "filename", "docType", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer", "nullable": true },
          "l2NodeId": { "type": "integer" },
          "filename": { "type": "string" },
          "docType": {
            "type": "string",
            "enum": ["markdown", "txt", "pdf", "docx", "pptx", "build_artifact"]
          },
          "content": { "type": "string" },
          "contentHash": {
            "type": "string",
            "description": "SHA-256 hash of raw content",
            "nullable": true
          },
          "commitSha": {
            "type": "string",
            "description": "Optional commit SHA to associate with this document",
            "nullable": true
          },
          "affiliatedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp when document was associated with a project",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["unaffiliated", "affiliated"],
            "default": "unaffiliated"
          },
          "createdAt": { "type": "string" }
        }
      },
      "DocumentIngestInput": {
        "type": "object",
        "required": ["filename", "content"],
        "properties": {
          "filename": { "type": "string", "minLength": 1 },
          "content": { "type": "string", "minLength": 1 },
          "docType": {
            "type": "string",
            "enum": ["markdown", "txt", "pdf", "docx", "pptx", "build_artifact"]
          },
          "commitSha": {
            "type": "string",
            "description": "Optional commit SHA to associate with this document"
          }
        }
      },
      "GitIngestInput": {
        "type": "object",
        "properties": {
          "repoUrl": { "type": "string" },
          "branch": { "type": "string" },
          "limit": { "type": "integer", "default": 100 },
          "githubToken": { "type": "string" },
          "mode": { "type": "string", "enum": ["full", "incremental"], "default": "full" }
        }
      },
      "GitIngestResult": {
        "type": "object",
        "required": ["commitsIngested", "commitsSkipped", "totalFetched"],
        "properties": {
          "commitsIngested": { "type": "integer" },
          "commitsSkipped": { "type": "integer" },
          "totalFetched": { "type": "integer" }
        }
      },
      "SvnIngestInput": {
        "type": "object",
        "required": ["svnUrl"],
        "properties": {
          "svnUrl": {
            "type": "string",
            "description": "SVN repository URL (e.g. svn+ssh://... or https://...)"
          },
          "username": { "type": "string", "description": "SVN username (optional)" },
          "password": { "type": "string", "description": "SVN password (optional)" },
          "startRevision": {
            "type": "integer",
            "description": "Starting revision number (optional, defaults to 1)"
          },
          "endRevision": {
            "oneOf": [{ "type": "integer" }, { "type": "string", "enum": ["HEAD"] }],
            "description": "Ending revision number (optional, defaults to HEAD)"
          },
          "mode": { "type": "string", "enum": ["full", "incremental"], "default": "full" }
        }
      },
      "SvnIngestResult": {
        "type": "object",
        "required": ["ingested", "skipped", "errors"],
        "properties": {
          "ingested": { "type": "integer" },
          "skipped": { "type": "integer" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "IngestStatusResponse": {
        "type": "object",
        "required": ["projectId", "vcsType", "pendingCommits"],
        "properties": {
          "projectId": { "type": "string" },
          "vcsType": { "type": "string", "enum": ["git", "svn"] },
          "lastGitIngestedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp of last Git ingest, or null",
            "nullable": true
          },
          "lastSvnRevision": {
            "type": "integer",
            "description": "Last ingested SVN revision number, or null",
            "nullable": true
          },
          "pendingCommits": {
            "type": "integer",
            "description": "Number of commits not yet processed"
          }
        }
      },
...
    }
  }
}